Crate tagmanager1

Features

Handle the following Resources with ease from the central hub ...

accounts (containers.create, containers.delete, containers.get, containers.list, containers.macros.create, containers.macros.delete, containers.macros.get, containers.macros.list, containers.macros.update, containers.rules.create, containers.rules.delete, containers.rules.get, containers.rules.list, containers.rules.update, containers.tags.create, containers.tags.delete, containers.tags.get, containers.tags.list, containers.tags.update, containers.triggers.create, containers.triggers.delete, containers.triggers.get, containers.triggers.list, containers.triggers.update, containers.update, containers.variables.create, containers.variables.delete, containers.variables.get, containers.variables.list, containers.variables.update, containers.versions.create, containers.versions.delete, containers.versions.get, containers.versions.list, containers.versions.publish, containers.versions.restore, containers.versions.undelete, containers.versions.update, get, list, permissions.create, permissions.delete, permissions.get, permissions.list, permissions.update and update)

Everything else about the Tag Manager API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.accounts().containers_list(...).doit()
let r = hub.accounts().permissions_list(...).doit()
let r = hub.accounts().containers_versions_undelete(...).doit()
let r = hub.accounts().permissions_create(...).doit()
let r = hub.accounts().permissions_delete(...).doit()
let r = hub.accounts().containers_get(...).doit()
let r = hub.accounts().containers_versions_list(...).doit()
let r = hub.accounts().containers_triggers_update(...).doit()
let r = hub.accounts().containers_triggers_get(...).doit()
let r = hub.accounts().containers_delete(...).doit()
let r = hub.accounts().containers_create(...).doit()
let r = hub.accounts().containers_tags_delete(...).doit()
let r = hub.accounts().containers_rules_update(...).doit()
let r = hub.accounts().containers_rules_delete(...).doit()
let r = hub.accounts().containers_tags_list(...).doit()
let r = hub.accounts().containers_versions_publish(...).doit()
let r = hub.accounts().containers_tags_create(...).doit()
let r = hub.accounts().containers_triggers_list(...).doit()
let r = hub.accounts().containers_versions_delete(...).doit()
let r = hub.accounts().update(...).doit()
let r = hub.accounts().containers_macros_delete(...).doit()
let r = hub.accounts().containers_versions_create(...).doit()
let r = hub.accounts().permissions_get(...).doit()
let r = hub.accounts().containers_rules_create(...).doit()
let r = hub.accounts().containers_versions_restore(...).doit()
let r = hub.accounts().containers_rules_get(...).doit()
let r = hub.accounts().containers_variables_create(...).doit()
let r = hub.accounts().containers_variables_list(...).doit()
let r = hub.accounts().containers_macros_create(...).doit()
let r = hub.accounts().containers_tags_get(...).doit()
let r = hub.accounts().containers_variables_get(...).doit()
let r = hub.accounts().containers_triggers_delete(...).doit()
let r = hub.accounts().containers_macros_list(...).doit()
let r = hub.accounts().containers_triggers_create(...).doit()
let r = hub.accounts().containers_macros_update(...).doit()
let r = hub.accounts().list(...).doit()
let r = hub.accounts().permissions_update(...).doit()
let r = hub.accounts().containers_variables_delete(...).doit()
let r = hub.accounts().get(...).doit()
let r = hub.accounts().containers_update(...).doit()
let r = hub.accounts().containers_rules_list(...).doit()
let r = hub.accounts().containers_tags_update(...).doit()
let r = hub.accounts().containers_macros_get(...).doit()
let r = hub.accounts().containers_versions_update(...).doit()
let r = hub.accounts().containers_variables_update(...).doit()
let r = hub.accounts().containers_versions_get(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
tagmanager1 = "0.0.1"

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate tagmanager1;
use tagmanager1::Trigger;
use tagmanager1::cmn::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use tagmanager1::TagManager;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = TagManager::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req: Trigger = Default::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.accounts().containers_triggers_update(&req, "accountId", "containerId", "triggerId")
             .fingerprint("accusam")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(value) => println!("Result Value: {:?}", value),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Modules

cmn

Macros

map!
This macro is advertised in the documentation, which is why we deliver it as well

Structs

Account
Represents a Google Tag Manager Account.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

containers.list accounts (none)

permissions.list accounts (none)

containers.versions.undelete accounts (none)

permissions.create accounts (none)

permissions.delete accounts (none)

containers.get accounts (none)

containers.versions.list accounts (none)

containers.triggers.update accounts (none)

containers.triggers.get accounts (none)

containers.delete accounts (none)

containers.create accounts (none)

containers.tags.delete accounts (none)

containers.rules.update accounts (none)

containers.rules.delete accounts (none)

containers.tags.list accounts (none)

containers.versions.publish accounts (none)

containers.tags.create accounts (none)

containers.triggers.list accounts (none)

containers.versions.delete accounts (none)

update accounts (request|response)

containers.macros.delete accounts (none)

containers.versions.create accounts (none)

permissions.get accounts (none)

containers.rules.create accounts (none)

containers.versions.restore accounts (none)

containers.rules.get accounts (none)

containers.variables.create accounts (none)

containers.variables.list accounts (none)

containers.macros.create accounts (none)

containers.tags.get accounts (none)

containers.variables.get accounts (none)

containers.triggers.delete accounts (none)

containers.macros.list accounts (none)

containers.triggers.create accounts (none)

containers.macros.update accounts (none)

list accounts (none)

permissions.update accounts (none)

containers.variables.delete accounts (none)

get accounts (response)

containers.update accounts (none)

containers.rules.list accounts (none)

containers.tags.update accounts (none)

containers.macros.get accounts (none)

containers.versions.update accounts (none)

containers.variables.update accounts (none)

containers.versions.get accounts (none)

AccountAccess
Defines the Google Tag Manager Account access permissions.

This type is not used in any activity, and only used as part of another schema.

AccountContainerCreateMethodBuilder
Creates a Container.

AccountContainerDeleteMethodBuilder
Deletes a Container.

AccountContainerGetMethodBuilder
Gets a Container.

AccountContainerListMethodBuilder
Lists all Containers that belongs to a GTM Account.

AccountContainerMacroCreateMethodBuilder
Creates a GTM Macro.

AccountContainerMacroDeleteMethodBuilder
Deletes a GTM Macro.

AccountContainerMacroGetMethodBuilder
Gets a GTM Macro.

AccountContainerMacroListMethodBuilder
Lists all GTM Macros of a Container.

AccountContainerMacroUpdateMethodBuilder
Updates a GTM Macro.

AccountContainerRuleCreateMethodBuilder
Creates a GTM Rule.

AccountContainerRuleDeleteMethodBuilder
Deletes a GTM Rule.

AccountContainerRuleGetMethodBuilder
Gets a GTM Rule.

AccountContainerRuleListMethodBuilder
Lists all GTM Rules of a Container.

AccountContainerRuleUpdateMethodBuilder
Updates a GTM Rule.

AccountContainerTagCreateMethodBuilder
Creates a GTM Tag.

AccountContainerTagDeleteMethodBuilder
Deletes a GTM Tag.

AccountContainerTagGetMethodBuilder
Gets a GTM Tag.

AccountContainerTagListMethodBuilder
Lists all GTM Tags of a Container.

AccountContainerTagUpdateMethodBuilder
Updates a GTM Tag.

AccountContainerTriggerCreateMethodBuilder
Creates a GTM Trigger.

AccountContainerTriggerDeleteMethodBuilder
Deletes a GTM Trigger.

AccountContainerTriggerGetMethodBuilder
Gets a GTM Trigger.

AccountContainerTriggerListMethodBuilder
Lists all GTM Triggers of a Container.

AccountContainerTriggerUpdateMethodBuilder
Updates a GTM Trigger.

AccountContainerUpdateMethodBuilder
Updates a Container.

AccountContainerVariableCreateMethodBuilder
Creates a GTM Variable.

AccountContainerVariableDeleteMethodBuilder
Deletes a GTM Variable.

AccountContainerVariableGetMethodBuilder
Gets a GTM Variable.

AccountContainerVariableListMethodBuilder
Lists all GTM Variables of a Container.

AccountContainerVariableUpdateMethodBuilder
Updates a GTM Variable.

AccountContainerVersionCreateMethodBuilder
Creates a Container Version.

AccountContainerVersionDeleteMethodBuilder
Deletes a Container Version.

AccountContainerVersionGetMethodBuilder
Gets a Container Version.

AccountContainerVersionListMethodBuilder
Lists all Container Versions of a GTM Container.

AccountContainerVersionPublishMethodBuilder
Publishes a Container Version.

AccountContainerVersionRestoreMethodBuilder
Restores a Container Version. This will overwrite the container's current configuration (including its macros, rules and tags). The operation will not have any effect on the version that is being served (i.e. the published version).

AccountContainerVersionUndeleteMethodBuilder
Undeletes a Container Version.

AccountContainerVersionUpdateMethodBuilder
Updates a Container Version.

AccountGetMethodBuilder
Gets a GTM Account.

AccountListMethodBuilder
Lists all GTM Accounts that a user has access to.

AccountMethodsBuilder
A builder providing access to all methods supported on account resources. It is not used directly, but through the TagManager hub.

AccountPermissionCreateMethodBuilder
Creates a user's Account & Container Permissions.

AccountPermissionDeleteMethodBuilder
Removes a user from the account, revoking access to it and all of its containers.

AccountPermissionGetMethodBuilder
Gets a user's Account & Container Permissions.

AccountPermissionListMethodBuilder
List all users that have access to the account along with Account and Container Permissions granted to each of them.

AccountPermissionUpdateMethodBuilder
Updates a user's Account & Container Permissions.

AccountUpdateMethodBuilder
Updates a GTM Account.

Condition
Represents a predicate.

This type is not used in any activity, and only used as part of another schema.

Container
Represents a Google Tag Manager Container.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

containers.update accounts (request|response)

containers.create accounts (request|response)

containers.get accounts (response)

ContainerAccess
Defines the Google Tag Manager Container access permissions.

This type is not used in any activity, and only used as part of another schema.

ContainerVersion
Represents a Google Tag Manager Container Version.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

containers.versions.update accounts (request|response)

containers.versions.get accounts (response)

containers.versions.undelete accounts (response)

containers.versions.restore accounts (response)

ContainerVersionHeader
Represents a Google Tag Manager Container Version Header.

This type is not used in any activity, and only used as part of another schema.

CreateContainerVersionRequestVersionOptions
Options for new container versions.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

containers.versions.create accounts (request)

CreateContainerVersionResponse
Create container versions response.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

containers.versions.create accounts (response)

ListAccountUsersResponse
List AccountUsers Response.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

permissions.list accounts (response)

ListAccountsResponse
List Accounts Response.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list accounts (response)

ListContainerVersionsResponse
List container versions response.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

containers.versions.list accounts (response)

ListContainersResponse
List Containers Response.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

containers.list accounts (response)

ListMacrosResponse
List Macros Response.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

containers.macros.list accounts (response)

ListRulesResponse
List Rules Response.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

containers.rules.list accounts (response)

ListTagsResponse
List Tags Response.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

containers.tags.list accounts (response)

ListTriggersResponse
List triggers response.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

containers.triggers.list accounts (response)

ListVariablesResponse
List Variables Response.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

containers.variables.list accounts (response)

Macro
Represents a Google Tag Manager Macro.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

containers.macros.create accounts (request|response)

containers.macros.get accounts (response)

containers.macros.update accounts (request|response)

Parameter
Represents a Google Tag Manager Parameter.

This type is not used in any activity, and only used as part of another schema.

PublishContainerVersionResponse
Publish container version response.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

containers.versions.publish accounts (response)

Rule
Represents a Google Tag Manager Rule.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

containers.rules.get accounts (response)

containers.rules.create accounts (request|response)

containers.rules.update accounts (request|response)

Tag
Represents a Google Tag Manager Tag.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

containers.tags.update accounts (request|response)

containers.tags.get accounts (response)

containers.tags.create accounts (request|response)

TagManager
Central instance to access all TagManager related resource activities

Trigger
Represents a Google Tag Manager Trigger

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

containers.triggers.update accounts (request|response)

containers.triggers.get accounts (response)

containers.triggers.create accounts (request|response)

UserAccess
Represents a user's permissions to an account and its container.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

permissions.get accounts (response)

permissions.create accounts (request|response)

permissions.update accounts (request|response)

Variable
Represents a Google Tag Manager Variable.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

containers.variables.create accounts (request|response)

containers.variables.update accounts (request|response)

containers.variables.get accounts (response)

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.